+++ Cinder.sln
@@ -19,13 +19,27 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Api", "Api", "{6364676A-C88
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application", "src\Api\Application\Application.csproj", "{6F9A6777-6975-4C4F-8BFB-4A285B14224C}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Indexer", "Indexer", "{5B2F61D1-8210-4150-BEA1-7962533F6D2C}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Indexers", "Indexers", "{5B2F61D1-8210-4150-BEA1-7962533F6D2C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Host", "src\Indexer\Host\Host.csproj", "{9F204F17-218D-4505-B70A-06298F683AF1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Host", "src\Indexers\BlockIndexer\Host\Host.csproj", "{9F204F17-218D-4505-B70A-06298F683AF1}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{83EE16C6-077E-427B-B56D-B6E92305C489}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Search", "src\Api\Search\Search.csproj", "{165C2F65-224E-4A69-86D2-455DE2B7AB6C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Search", "src\Api\Search\Search.csproj", "{165C2F65-224E-4A69-86D2-455DE2B7AB6C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BlockIndexer", "BlockIndexer", "{145E8210-2743-4803-B216-AAF6C2572AF1}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Host", "src\Indexers\AddressIndexer\Host\Host.csproj", "{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AddressIndexer", "AddressIndexer", "{E5B1DEF7-9C69-4179-9E56-ED072955FE2A}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "StatsIndexer", "StatsIndexer", "{FD854B5C-A28F-4CC7-9DC1-669A7484754C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Host", "src\Indexers\StatsIndexer\Host\Host.csproj", "{489B4E33-73D8-4CC2-9367-89DE5990E69C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{6BC991C3-06FB-43DF-97FD-0F6D7E248FED}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HostBase", "src\Indexers\Shared\HostBase\HostBase\HostBase.csproj", "{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -133,6 +147,42 @@ Global
 		{165C2F65-224E-4A69-86D2-455DE2B7AB6C}.Release|x64.Build.0 = Release|Any CPU
 		{165C2F65-224E-4A69-86D2-455DE2B7AB6C}.Release|x86.ActiveCfg = Release|Any CPU
 		{165C2F65-224E-4A69-86D2-455DE2B7AB6C}.Release|x86.Build.0 = Release|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Debug|x64.Build.0 = Debug|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Debug|x86.Build.0 = Debug|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Release|x64.ActiveCfg = Release|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Release|x64.Build.0 = Release|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Release|x86.ActiveCfg = Release|Any CPU
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D}.Release|x86.Build.0 = Release|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Debug|x64.Build.0 = Debug|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Debug|x86.Build.0 = Debug|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Release|x64.ActiveCfg = Release|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Release|x64.Build.0 = Release|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Release|x86.ActiveCfg = Release|Any CPU
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C}.Release|x86.Build.0 = Release|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Debug|x64.Build.0 = Debug|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Debug|x86.Build.0 = Debug|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Release|x64.ActiveCfg = Release|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Release|x64.Build.0 = Release|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Release|x86.ActiveCfg = Release|Any CPU
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -146,9 +196,16 @@ Global
 		{6364676A-C88D-4966-8E7F-1E130F7A230A} = {E3B590F2-0BE3-44A3-A864-2B67EAA9467A}
 		{6F9A6777-6975-4C4F-8BFB-4A285B14224C} = {6364676A-C88D-4966-8E7F-1E130F7A230A}
 		{5B2F61D1-8210-4150-BEA1-7962533F6D2C} = {E3B590F2-0BE3-44A3-A864-2B67EAA9467A}
-		{9F204F17-218D-4505-B70A-06298F683AF1} = {5B2F61D1-8210-4150-BEA1-7962533F6D2C}
+		{9F204F17-218D-4505-B70A-06298F683AF1} = {145E8210-2743-4803-B216-AAF6C2572AF1}
 		{83EE16C6-077E-427B-B56D-B6E92305C489} = {E3B590F2-0BE3-44A3-A864-2B67EAA9467A}
 		{165C2F65-224E-4A69-86D2-455DE2B7AB6C} = {6364676A-C88D-4966-8E7F-1E130F7A230A}
+		{145E8210-2743-4803-B216-AAF6C2572AF1} = {5B2F61D1-8210-4150-BEA1-7962533F6D2C}
+		{AB76F2BD-9C2A-4D6A-AD39-4C742CD60B3D} = {E5B1DEF7-9C69-4179-9E56-ED072955FE2A}
+		{E5B1DEF7-9C69-4179-9E56-ED072955FE2A} = {5B2F61D1-8210-4150-BEA1-7962533F6D2C}
+		{FD854B5C-A28F-4CC7-9DC1-669A7484754C} = {5B2F61D1-8210-4150-BEA1-7962533F6D2C}
+		{489B4E33-73D8-4CC2-9367-89DE5990E69C} = {FD854B5C-A28F-4CC7-9DC1-669A7484754C}
+		{6BC991C3-06FB-43DF-97FD-0F6D7E248FED} = {5B2F61D1-8210-4150-BEA1-7962533F6D2C}
+		{15276B8A-1D05-4AD1-A054-E0BF4C6707DE} = {6BC991C3-06FB-43DF-97FD-0F6D7E248FED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B8C3070B-22C5-4AD6-AE66-DC081781EDCD}
