--- src/Indexer/Host/Infrastructure/IAddressRefresherRunner.cs
@@ -1,10 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-
-namespace Cinder.Indexer.Host.Infrastructure
-{
-    public interface IAddressRefresherRunner
-    {
-        Task Run(CancellationToken cancellationToken = default);
-    }
-}
