+++ src/Indexers/BlockIndexer/Host/Infrastructure/Program.cs
@@ -1,0 +1,40 @@
+using System.IO;
+using System.Threading.Tasks;
+using Cinder.Extensions;
+using Cinder.Extensions.DependencyInjection;
+using Cinder.Indexers.BlockIndexer.Host.Infrastructure.Hosting;
+using Cinder.Indexers.HostBase;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
+namespace Cinder.Indexers.BlockIndexer.Host.Infrastructure
+{
+    public class Program
+    {
+        public static IConfiguration Configuration { get; } = new ConfigurationBuilder()
+            .SetBasePath(Directory.GetCurrentDirectory())
+            .AddJsonFile("appsettings.json", false, true)
+            .AddEnvironmentVariables()
+            .Build();
+
+        public static async Task<int> Main(string[] args)
+        {
+            HostWrapper hostWrapper = new HostWrapper
+            {
+                DefaultBuilder = new HostBuilder().ConfigureServices((hostContext, services) =>
+                {
+                    services.Configure<ConsoleLifetimeOptions>(options => options.SuppressStatusMessages = true);
+                    services.AddHostedService<BlockIndexerService>();
+                    services.Configure<SettingsBase>(options => Configuration.Bind(options));
+                    services.AddBlockchain();
+                    services.AddDatabase();
+                    services.AddEvents();
+                    services.AddWeb3();
+                })
+            };
+
+            return await hostWrapper.Run("Block Indexer", Configuration).AnyContext();
+        }
+    }
+}
