+++ src/Indexers/BlockIndexer/Host/appsettings.json
@@ -1,0 +1,31 @@
+{
+  "Serilog": {
+    "MinimumLevel": {
+      "Default": "Debug",
+      "Override": {
+        "Microsoft": "Warning",
+        "System": "Warning"
+      }
+    },
+    "WriteTo": [
+      {
+        "Name": "Async",
+        "Args": {
+          "configure": [
+            {
+              "Name": "Console"
+            }
+          ]
+        }
+      }
+    ]
+  },
+  "Database": {
+    "ConnectionString": "mongodb://localhost:27017",
+    "Tag": "ellaism",
+    "Locale": "en"
+  },
+  "Node": {
+    "RpcUrl": "http://localhost:8545"
+  }
+}
