+++ src/Indexers/Shared/HostBase/HostBase/Extensions/DependencyInjection/Event.cs
