+++ src/Indexers/Shared/HostBase/HostBase/HostBase.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>Cinder.Indexers.HostBase</AssemblyName>
+    <RootNamespace>Cinder.Indexers.HostBase</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Foundatio" Version="8.1.2126" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.0" />
+    <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.0" />
+    <PackageReference Include="Nethereum.Parity" Version="3.4.0" />
+    <PackageReference Include="Serilog.Extensions.Hosting" Version="3.0.0" />
+    <PackageReference Include="Serilog.Settings.Configuration" Version="3.1.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\Shared\Core\Core.csproj" />
+    <ProjectReference Include="..\..\..\..\Shared\Data\Data.csproj" />
+  </ItemGroup>
+
+</Project>
